fix(backend): itsm的处理人关联审批人 #8688

+++ dbm-ui/backend/ticket/builders/redis/redis_toolbox_redis_scale_updown.py
@@ -44,7 +44,7 @@ class RedisScaleUpDownDetailSerializer(SkipToRepresentationMixin, serializers.Se
         shard_num = serializers.IntegerField(help_text=_("集群分片数"))
         group_num = serializers.IntegerField(help_text=_("部署机器组数"))
         db_version = serializers.CharField(help_text=_("版本号"))
-        capacity = serializers.IntegerField(help_text=_("当前容量需求"))
+        capacity = serializers.FloatField(help_text=_("当前容量需求"))
         future_capacity = serializers.IntegerField(help_text=_("未来容量需求"))
         online_switch_type = serializers.ChoiceField(
             help_text=_("切换类型"), choices=SwitchConfirmType.get_choices(), default=SwitchConfirmType.NO_CONFIRM
